[scanner] refactor: extract drift cache and gitops SSE/operator modules
